Set first token in popular list as opposite token
